rag_backend/requirements.txt
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.6
fastapi==0.115.4
langchain==0.3.4
langchain_community==0.3.3
langchain_core==0.3.13
langchain_openai==0.2.4
langchain_pinecone==0.2.0
langchain_text_splitters==0.3.0
pinecone==5.3.1
uvicorn==0.32.0
